+++ backend/src/main/java/ca/metricalsky/winston/mappers/api/AuthorMapper.java
@@ -3,6 +3,7 @@ package ca.metricalsky.winston.mappers.api;
 import ca.metricalsky.winston.api.model.Author;
 import ca.metricalsky.winston.entity.AuthorEntity;
 import ca.metricalsky.winston.entity.view.AuthorDetailsView;
+import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Named;
@@ -18,13 +19,20 @@ public abstract class AuthorMapper {
 
     @Mapping(target = "handle", source = ".", qualifiedByName = "mapHandle")
     @Mapping(target = "profileImageUrl", source = ".", qualifiedByName = "mapProfileImageUrl")
-    @Mapping(target = "statistics", ignore = true)
+    @Mapping(target = "authorStatistics", ignore = true)
+    @Mapping(target = "videoStatistics", ignore = true)
     public abstract Author toAuthor(AuthorEntity authorEntity);
+
+    @InheritInverseConfiguration
+    @Mapping(target = "displayName", source = "handle")
+    @Mapping(target = "lastFetchedAt", ignore = true)
+    public abstract AuthorEntity toAuthorEntity(Author author);
 
     @Mapping(target = ".", source = "author")
     @Mapping(target = "handle", source = "author", qualifiedByName = "mapHandle")
     @Mapping(target = "profileImageUrl", source = "author", qualifiedByName = "mapProfileImageUrl")
-    @Mapping(target = "statistics", source = ".")
+    @Mapping(target = "authorStatistics", source = ".")
+    @Mapping(target = "videoStatistics", ignore = true)
     public abstract Author toAuthor(AuthorDetailsView authorDetailsView);
 
     @Named("mapHandle")