+++ frontend/src/routes/channels/id/VideoCard.tsx
@@ -1,17 +1,12 @@
 import { Card, Col, Ratio, Row } from "react-bootstrap";
-import { ArrowUpLeftCircleFill } from "react-bootstrap-icons";
 import { Link } from "react-router";
 import { CommentCounts } from "#/components/comments/CommentCounts";
-import { Date } from "#/components/Date";
+import { VideoPublishedAt } from "#/components/videos/VideoPublishedAt";
 import { VideoCardSelectionCheckbox } from "#/routes/channels/id/VideoCardSelectionCheckbox";
 import { routes } from "#/utils/links";
-import type { Video } from "#/api";
+import type { VideoProps } from "#/types";
 
-type Props = {
-    video: Video;
-};
-
-export const VideoCard = ({ video }: Props) => (
+export const VideoCard = ({ video }: VideoProps) => (
     <Col className={"g-2"}>
         <Card className={"h-100"}>
             <VideoCardSelectionCheckbox video={video} />
@@ -39,9 +34,8 @@ export const VideoCard = ({ video }: Props) => (
                     />
                 </Row>
             </Card.Footer>
-            <Card.Footer className={"d-flex align-items-center"}>
-                <ArrowUpLeftCircleFill className={"me-2"} />
-                <Date date={video.published_at} />
+            <Card.Footer>
+                <VideoPublishedAt video={video} />
             </Card.Footer>
         </Card>
     </Col>